# Station query optimization