--- a/.github/workflows/lint.yml
+++ b/.github/workflows/lint.yml
@@ -17,7 +17,6 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: |
-            "2.7"
             "3.10"
 
       - name: Install python requirements