--- a/.github/workflows/lint.yml
+++ b/.github/workflows/lint.yml
@@ -16,9 +16,7 @@ jobs:
       - name: Setup Python 3.10
         uses: actions/setup-python@v4
         with:
-          python-version: |
-            2.7
-            3.10
+          python-version: 3.10
 
       - name: Install python requirements
         run: |