--- a/.github/workflows/lint.yml
+++ b/.github/workflows/lint.yml
@@ -21,7 +21,7 @@ jobs:
       - name: Install python requirements
         run: |
           pip install -U pip
-          pip install poetry
+          pip install poetry==1.6.1
           poetry install
         env:
           POETRY_VIRTUALENVS_CREATE: false