--- a/.github/workflows/lint.yml
+++ b/.github/workflows/lint.yml
@@ -21,6 +21,7 @@ jobs:
       - name: Install python requirements
         run: |
           pip install -U pip
+          pip install virtualenv=20.17.1
           pip install poetry==1.6.1
           poetry install
         env: