--- a/.github/workflows/lint.yml
+++ b/.github/workflows/lint.yml
@@ -21,8 +21,7 @@ jobs:
       - name: Install python requirements
         run: |
           pip install -U pip
-          pip install poetry==1.4.2
-          pip install virtualenv==20.17.1
+          pip install poetry
           poetry install
         env:
           POETRY_VIRTUALENVS_CREATE: false