The cause of slow response time on the `stations/search` endpoint is the join with `stations_species`. When there is no filter on the `species_id` column, up to ~400 rows on `stations_species` are scanned and ~400 identical rows are generated for each station (which are then deduped by the ORM so it's not quite obvious from the outside). When there is a filter on `species_id`, the join is not expensive because Postgres executes the filter before the join. 

With this fix, on my laptop:
- Filter on `date`: up to ~4s $\longrightarrow$ up to ~80ms
- Filter on `fao_area`: ~200ms $\longrightarrow$ ~10ms
- Any filter involving `species_id`: ~10ms $\longrightarrow$ ~10ms